Time: 7 ms (27.04%), Space: 5.8 MB (76.09%) - LeetHub

+++ 29-divide-two-integers/29-divide-two-integers.cpp
@@ -6,10 +6,10 @@ public:
     
         int sign =(dividend>0 ^ divisor>0)?0:1;
       
-        unsigned int ans=0;
+        long int ans=0;
         
-        unsigned int a=abs(dividend);
-        unsigned int b=abs(divisor);
+        long int a=abs(dividend);
+        long int b=abs(divisor);
         
          if(b>a) return 0;
         
@@ -24,17 +24,7 @@ public:
             a=a-(b<<k);
             
         }  
-        // while(a >= b){  // while dividend is greater than or equal to divisor
-        //     short q = 0;
-        //     while(a > (b<<(q+1)))
-        //         q++;
-        //     ans += (1<<q);  // add the power of 2 found to the answer
-        //     a = a - (b<<q);  // reduce the dividend by divisor * power of 2 found
-        // }
-        
-        
-        
-        
+
          if(ans>INT_MAX)
         {
            return sign==1 ? INT_MAX:INT_MIN;
